+++ codes.go
@@ -1,0 +1,7 @@
+package health
+
+// The health check service is now online and ready for service.
+const MServiceAvailable = "HEALTH_AVAILABLE"
+
+// A request to the health check endpoint failed.
+const ERequestFailed = "HEALTH_REQUEST_FAILED"