+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # Changelog
 
-## 1.0.1
+## 1.1.0: Added health check client
+
+This release adds a dedicated client library for health checks. It also adds an "Enable" field to the configuration.
+
+## 1.0.1: JSON and YAML tags
 
 Add JSON and YAML tag as well as a sane default for the HTTP configuration.
 
-## 1.0.0
+## 1.0.0: Initial version
 
 Initial version of a health check service using an external configuration.